requirements.txt
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.0.0b6
marshmallow-sqlalchemy==0.23.1
mysql-connector-python==8.0.21
mysqlclient @ file:///C:/Users/workspace/desafios/mercafacil/mysqlclient-1.4.6-cp38-cp38-win32.whl
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
